Move the iter methods back to the root for doc order
